# api/swagger.yaml
basePath: /
definitions:
  structures.Authentication:
    properties:
      token:
        type: string
      totp:
        type: string
    type: object
  structures.TOTPToken:
    properties:
      TOTP:
        type: string
    type: object
  structures.UserAsk:
    properties:
      username:
        type: string
    type: object
  structures.UserChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      username:
        type: string
    type: object
  structures.UserCreateTOTP:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  structures.UserReinitialize:
    properties:
      authentication:
        $ref: '#/definitions/structures.Authentication'
      new_password:
        type: string
      username:
        type: string
    type: object
  structures.UserUnlock:
    properties:
      authentication:
        $ref: '#/definitions/structures.Authentication'
      username:
        type: string
    type: object
  structures.UserVerifyTOTP:
    properties:
      totp:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: Cloud dev team
    url: https://github.com/be-ys-cloud
  description: DORY (stands for Domain Operator - Recover by Yourself) is a tool who
    enable user to unlock their AD account, and change/reinitialize password. This
    API is NOT intented to be RESTful !
  license:
    name: AGPL
    url: https://www.gnu.org/licenses/agpl-3.0.fr.html
  title: Dory-API
  version: 1.0.0
paths:
  /change_password:
    post:
      description: Change a user's password.
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structures.UserChangePassword'
      responses:
        "200":
          description: OK - Mail changed
        "400":
          description: Invalid payload
        "500":
          description: An error occured.
      security:
      - BasicAuth: []
      summary: Change a user's password.
      tags:
      - change_password
  /reinitialize:
    post:
      description: Reinitialize a user's password.
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structures.UserReinitialize'
      responses:
        "200":
          description: OK - Password changed
        "400":
          description: Missing data in payload
        "500":
          description: An error occured.
      security:
      - BasicAuth: []
      summary: Reinitialize a user's password.
      tags:
      - reinitialization
  /request/{kind}:
    post:
      description: Ask server to send email with a link to reset an account password,
        or unlock it.
      parameters:
      - description: 'Kind of request : reinitialize or unlock.'
        in: path
        name: kind
        required: true
        type: string
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structures.UserAsk'
      responses:
        "200":
          description: OK - Check your mailbox
        "400":
          description: Missing username in payload
        "500":
          description: An error occurred.
      security:
      - BasicAuth: []
      summary: Ask server to send email with a link to reset an account password,
        or unlock it.
      tags:
      - demand
  /totp/create:
    post:
      description: Verify a TOTP for this user.
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structures.UserVerifyTOTP'
      responses:
        "200":
          description: TOTP valid
        "400":
          description: Invalid payload
        "401":
          description: TOTP invalid
        "500":
          description: An error occured.
      security:
      - BasicAuth: []
      summary: Verify a TOTP token for this user.
      tags:
      - totp
  /unlock:
    post:
      description: Unlock a user.
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structures.UserUnlock'
      responses:
        "200":
          description: OK - User unlocked
        "400":
          description: Missing data in payload
        "500":
          description: An error occured.
      security:
      - BasicAuth: []
      summary: Unlock a user.
      tags:
      - unlock
schemes:
- https
swagger: "2.0"